backend/requirements.txt
fastapi
uvicorn
torch
transformers
pillow
python-multipart
easyocr
opencv-python-headless

